+++ packages/hooks/src/next/useOrderEntry/useOrderStore.ts
@@ -22,6 +22,11 @@ const initialOrderState = {
   tp_offset: "",
   sl_offset: "",
   total: "",
+  min_price: "",
+  max_price: "",
+  totalOrders: "",
+  distribution_type: "",
+  skew: "",
   // symbol: "",
 };
 
@@ -29,7 +34,7 @@ export const useOrderStore = (initialOrder: OrderEntryStateEntity) => {
   const [entry, setEntry] = useState<OrderEntryStateEntity>(initialOrder);
   const [estLeverage, setEstLeverage] = useState<number | null>(null);
   const [estLiquidationPrice, setEstLiquidationPrice] = useState<number | null>(
-    null
+    null,
   );
   const [errors, setErrors] = useState<
     Partial<Record<keyof FullOrderState, string>>
@@ -40,19 +45,19 @@ export const useOrderStore = (initialOrder: OrderEntryStateEntity) => {
     setEntry(
       produce((draft) => {
         return { ...draft, ...order };
-      })
+      }),
       // (prev) => ({ ...prev, ...order })
     );
   };
 
   const updateOrderByKey = <K extends keyof FullOrderState>(
     key: K,
-    value: FullOrderState[K]
+    value: FullOrderState[K],
   ) => {
     setEntry(
       produce((draft) => {
         draft[key] = value;
-      })
+      }),
     );
   };
 
@@ -73,7 +78,7 @@ export const useOrderStore = (initialOrder: OrderEntryStateEntity) => {
       produce((draft) => ({
         ...draft,
         ...(order ?? initialOrderState),
-      }))
+      })),
     );
   };