+++ packages/hooks/src/orderlyContext.ts
@@ -1,13 +1,11 @@
 import { createContext } from "react";
-
 import {
   type ConfigStore,
   type OrderlyKeyStore,
   type getWalletAdapterFunc,
-  WalletAdapter
+  WalletAdapter,
 } from "@orderly.network/core";
-
-import { Chain, NetworkId } from "@orderly.network/types";
+import { API, Chain, NetworkId } from "@orderly.network/types";
 import { Chains } from "./orderly/useChains";
 
 export type filteredChains = {
@@ -36,6 +34,15 @@ export interface OrderlyConfigContextState {
   // extraApis:ExtraAPIs
   filteredChains?: filteredChains | null;
   customChains?: Chains<undefined, undefined>;
+  chainTransformer?: (params: {
+    chains: API.Chain[];
+    tokenChains: API.Chain[];
+    chainInfos: any[];
+    swapChains: any[];
+    mainnet: boolean;
+  }) => API.Chain[];
+  /** enable swap deposit, default is true */
+  enableSwapDeposit?: boolean;
 }
 
 export const OrderlyContext = createContext<OrderlyConfigContextState>({