+++ packages/ui-scaffold/src/index.ts
@@ -34,3 +34,4 @@ export {
 export { MainLogo } from "./components/main/mainLogo";
 export * from "./components/languageSwitcher";
 export * from "./components/scanQRCode";
+export * from "./components/leftNav";