+++ packages/ui-transfer/src/components/depositForm/hooks/index.ts
@@ -2,5 +2,3 @@ export { useActionType } from "./useActionType";
 export { useChainSelect, type CurrentChain } from "./useChainSelect";
 export { useDepositAction } from "./useDepositAction";
 export { useInputStatus } from "./useInputStatus";
-export { useToken } from "./useToken";
-export { useToken as useToken_v2 } from "./useToken_v2";