+++ packages/ui-transfer/src/components/withdrawForm/withdrawForm.ui.tsx
@@ -31,7 +31,7 @@ export const WithdrawForm: React.FC<WithdrawFormProps> = (props) => {
     disabled,
     quantity,
     onQuantityChange,
-    token,
+    sourceToken,
     amount,
     maxQuantity,
     chains,
@@ -60,7 +60,7 @@ export const WithdrawForm: React.FC<WithdrawFormProps> = (props) => {
           <QuantityInput
             value={quantity}
             onValueChange={onQuantityChange}
-            token={token}
+            token={sourceToken}
             tokens={sourceTokens}
             onTokenChange={onSourceTokenChange}
             status={props.inputStatus}
@@ -71,7 +71,7 @@ export const WithdrawForm: React.FC<WithdrawFormProps> = (props) => {
           />
         </Box>
         <AvailableQuantity
-          token={token}
+          token={sourceToken}
           amount={amount}
           maxQuantity={maxQuantity.toString()}
           loading={props.balanceRevalidating}
@@ -115,7 +115,7 @@ export const WithdrawForm: React.FC<WithdrawFormProps> = (props) => {
               classNames={{
                 root: "oui-mt-[2px] oui-rounded-t-sm oui-rounded-b-xl",
               }}
-              token={token}
+              token={sourceToken}
               value={props.showQty}
               readOnly
             />