+++ packages/ui/src/icon/emptyData.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 
-function EmptyStateIcon() {
+export function EmptyStateIcon() {
   return (
     <svg
       xmlns="http://www.w3.org/2000/svg"